Renamed a duplicate building blocks markdown file (#235)
* Added Tenancy setup

Added the ability to run the lab through a tenancy with the appropriate setup information.

* Blurred sensitive info

Blurred sensitive info for ADB screens.

* Updates to add a Tenancy option to the lab

Modified the lab to add a tenancy option and fixed some sandbox screens to account for the differences from a tenacy.

* Renamed a building block file

Renamed the building block file connect-with-sql-worksheet-body.md since the lab uses a local version that wasn't being found. Now called adb-database-actions.md.

---------

Co-authored-by: Andy Rivenes <andy.rivenes@gmail.com>